dtypes nice and clean (#673)
* add dtype class

* dtypes

* buffers are lazy

* dtype is tracked by lazybuffer and GenericShape

* fix types in llvm

* llvm store

* dtype tests

* fix tests maybe

* fix flop counter

* fix CI

* CI fix and check format

* fix dtype and dtype check

* fix custom test

* fix test graph

+++ test/unit/test_graph.py
@@ -14,10 +14,10 @@ class TestGraph(unittest.TestCase):
     assert nx.is_isomorphic(G, RG, node_match=lambda x,y: x["label"] == y["label"], edge_match=lambda x,y: x["label"] == y["label"] if "label" in y else True)
 
   def test_add_graph(self):
-    a = CPUBuffer.fromCPU(np.ones((4,4)))
-    b = CPUBuffer.fromCPU(np.ones((4,4)))
+    a = CPUBuffer.fromCPU(np.ones((4,4), dtype=np.float32))
+    b = CPUBuffer.fromCPU(np.ones((4,4), dtype=np.float32))
     ast = LazyOp(BinaryOps.ADD, (a,b))
-    ret = CPUBuffer(np.ones((4,4)))
+    ret = CPUBuffer(np.ones((4,4), dtype=np.float32))
 
     RG = nx.DiGraph()
     RG.add_node(0, label="(4, 4)")
@@ -30,12 +30,12 @@ class TestGraph(unittest.TestCase):
     self.helper_compare_graph(RG)
 
   def test_add_sum_graph(self):
-    a = CPUBuffer.fromCPU(np.ones((4,4)))
-    b = CPUBuffer.fromCPU(np.ones((1,1)))
+    a = CPUBuffer.fromCPU(np.ones((4,4), dtype=np.float32))
+    b = CPUBuffer.fromCPU(np.ones((1,1), dtype=np.float32))
     op0 = LazyOp(MovementOps.RESHAPE, (b,), (4, 4))
     op1 = LazyOp(BinaryOps.ADD, (a,op0))
     ast = LazyOp(ReduceOps.SUM, (op1,), (1,1))
-    ret = CPUBuffer(np.ones((1,1)))
+    ret = CPUBuffer(np.ones((1,1), dtype=np.float32))
 
     RG = nx.DiGraph()
     RG.add_node(0, label="(4, 4)")
@@ -48,14 +48,14 @@ class TestGraph(unittest.TestCase):
     self.helper_compare_graph(RG)
 
   def test_add_graph_prune(self):
-    a = CPUBuffer.fromCPU(np.ones((1,1)))
+    a = CPUBuffer.fromCPU(np.ones((1,1), dtype=np.float32))
     ast = LazyOp(MovementOps.RESHAPE, (a,), (4, 4))
-    ret = CPUBuffer(np.ones((4,4)))
+    ret = CPUBuffer(np.ones((4,4), dtype=np.float32))
     log_op(ret, ast, show_graph=True)
 
-    b = CPUBuffer.fromCPU(np.ones((4,4)))
+    b = CPUBuffer.fromCPU(np.ones((4,4), dtype=np.float32))
     ast = LazyOp(BinaryOps.ADD, (ret,b))
-    ret = CPUBuffer(np.ones((4,4)))
+    ret = CPUBuffer(np.ones((4,4), dtype=np.float32))
     log_op(ret, ast, show_graph=True)
     prune_graph()
